chore: move ExpandableVideo to ui packages
