+++ packages/ui/src/components/ExpandableVideo/index.tsx
@@ -1,0 +1,1 @@
+export * from './ExpandableVideo'
